// packages/nex/src/language/ast/types.ts
/**
 * MIT License
 *
 * Copyright (c) 2025 Chris M. Perez
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

import type { Kind } from '../kinds/index.js';
import type { Location } from './location.js';
import type { NameNode } from './name.js';

export interface NamedTypeNode {
	readonly kind: typeof Kind.NAMED_TYPE;
	readonly name: NameNode;
	readonly loc?: Location | undefined;
}

export interface ListTypeNode {
	readonly kind: typeof Kind.LIST_TYPE;
	readonly type: TypeNode;
	readonly loc?: Location | undefined;
}

export interface NonNullTypeNode {
	readonly kind: typeof Kind.NON_NULL_TYPE;
	readonly type: TypeNode;
	readonly loc?: Location | undefined;
}

export interface OptionalTypeNode {
	readonly kind: typeof Kind.OPTIONAL_TYPE;
	readonly type: TypeNode;
	readonly loc?: Location | undefined;
}

export type TypeNode =
	| NamedTypeNode
	| ListTypeNode
	| NonNullTypeNode
	| OptionalTypeNode;